Add request pipelining TINKERPOP-1775/1774
This allows connections to be reused for different requests in parallel
which should lead to better utilization of connections / lower resource
usage. The ConnectionPool now also has a min and max size.
A new connection is only created if all existing connections reached
their max in-flight limit (which is also a new setting). If additionally
the max size of the ConnectionPool is reached, then an exception is
thrown which makes this a breaking change.